meson: calculate libpython_path with meson
The only place PY_LIB_LOC, PYTHON_VERSION, PY_ABI_FLAGS and PY_LIB_SUFFIX
are used are for creating the libpython_path. Instead of exposing all
of them, simply calculate the path with meson, and then export that.

Added bonus: using join_paths() the "/" will be in the right direction
for windows vs. linux as well.

Also for some installations of python on OSX (from macports as an example)
LIBPL is defined, so use that if it is.